Make cluster state version contain major and minor
This commit changes what is stored in the version storage
under cluster state directory. Version file represents
the version of the cluster state directory layout. Version
is now represented by two integers instead of one. They
represent major and minor version of the directory
layout format.

Minor version could be useful for the case when we need
to perform a backward-incompatible fix in version X but
version X + 1 already exists.