Improve versioning of application protocols
Application protocols in CC are catchup and raft. Both are versioned
and the version is negotiated for every connection. Before this
change, version was just a single integer number. Such a versioning
scheme could be a problem when a backward-incompatible fix is
required for version X but version X+1 already exists.

This commit changes the versioning scheme to be '<major>.<minor>'.
Both major and minor are non-negative integer values. New scheme
should allow us to perform a release of version 2.1 when a newer
protocol with version 3 exists.