Plan with minimum graph...
Plan_with_minimum_graph replaces the previous functionality
plan_with_minimum_cardinality, and solves the issue of getting plans on
an empty database (or supersmall) that do not scale as data is added.
The previous solution would set a minimum bound for operator cardinalities,
but this would not work if the operator had an input cardinality bigger
that the specified min cardinality. The new solution simply sets
minimum values for the graph statistics, which should provide a
whole-planner emulation of a bigger than actual graph, and therefore
give scalable plans always.