Better reuse of buffers in BlockBasedIndexPopulator
Buffers for writing scan updates are now associated with the population
flush threads, i.e. shared between all populating indexes in the
same job for that thread. This works because scan updates are
kept in an array and sorted where the buffer is briefly used when
writing that block. This makes number of scan update buffers depend
on number of workers only (which is capped to 8 a.t.m.), whereas
previously that number depended on number of workers AND number of
indexes in the population job.

Merge buffers are now reused between merge rounds and will therefore
depend only on merge factor. Previously each read block and each
written block would allocate a new buffer and so the number of buffers
depended on merge factor AND index size.

These changes put into some examples (on a 4-core machine):

- Populating 16 indexes, each index ~400k entries
  memory usage w/ old buffering: ~190 MiB
  memory usage w/ new buffering: ~41 MiB
- Populating 25 indexes, each index ~80mn entries
  memory usage w/ old buffering: ~2,6 GiB
  memory usage w/ new buffering: ~42 MiB

The old way of buffering didn't scale very well, but the new way
stays virtually flat. "Virtually" because there's one smaller buffer
allocated per populating index for writing external updates in the
new way of buffering.