Don't create a new LifeSupport for getting context
Creating a new life support just for getting the ExecutionContext will
basically start up a new database and create new thread pools etc.
For the morsel runtime this lead to the weird behaviour that we got
two `QueryManager` and we added queries to one and pulled from the
other, i.e. we never found any work to be done, In benchmarks we just
idly waited for any meaningful task to perform.