Replicator waits for leader a configurable amount of time
This could cause issues during shutdown or other times when a leader
was not available at all, potentially blocking for a very long time
or even forever.

A bunch of other fixes, smaller refactoring, comments, TODOs
where also added to make the code more understandable and make it
clear what remains to be looked at.

The main functional change is that we will wait for the configured
amount of time for a leader both at the very start of the replication
attempt as well as after replication attempts already have been made.

Previously we would give up an attempt immediately if a leader was
not available for the first attempt, but after a replication attempt
already had been attempted we would never give up waiting for a leader.
In fact we would be stuck forever in the wait() of the LeaderProvider
and never, ever leave the monitor not matter what happened in the
world outside!

The never giving up strategy is required if one is aiming for
exactly-once semantics, but the system is not really mature enough
for that yet.

The LeaderProvider also had some work done to fix a concurrency issue
(missing volatile on field) and to fold in more of the helper
functionality which the replicator requires.