Move SSL handler installation into NettyPipelineBuilder
SSL handler is installed in the pipeline only when SSL policy is
configured. Before this change, installation of SSL handlers
was done outside of NettyPipelineBuilder by pipeline wrappers.
This made it hard to keep the same SSL handler when
reinstalling a pipeline.

This commit removes pipeline wrappers and moves SSL handler
installation into NettyPipelineBuilder. Builder is now able
to keep the existing SSL handler and not create a new one.