Force serial pipelines to only have 1 continuation at any point...
Before we would schedule a full parallel AllNodeScan even if it was part
of a serial pipeline, meaning that only 1 thread could ever be used. To
make sure this does not happen again, I've made serial pipeline
continuations use a singleton buffer that throws if a second element is
added.