Wait for Raft membership after replacing Core in stress test
Previously we had a badly written MembershipWaiter in the startup
lifecycle and also only a single database to worry about.

The waiter was removed because it was badly written and also it
doesn't make all that much sense to wait for any particular
database. Databases now generally operate completely asynchronously
from each other and from the DBMS global part itself.