Start raft message logger in lifecycle
Starting means opening a file writer, which is closed when logger
is stopped. Before this change, a file writer was created in the
constructor. So logger was initialized even before any lifecycle
method was invoked. This means it wasn't stopped if some of the
preceding lifecycle instances failed to start.

Removed some unnecessary generics from logger method signatures.