Allow cached properties for entities in ref slots.
Slotted failed if the entity of a cached property was stored in a ref
slot. Now, we have a flag to distinguish whether the offset is for a
long or a ref slot. This will only minimally affect performance
for interpreted expressions. For compiled expressions, we decide which
code path to take at compile time.