Fix bug with CallingThreadQueryExecutor freeing cursors into two pools...
Before we used transaction.cursors as the backing CursorFactory for
query execution on CallingThreadQueryExecutor, but this would lead to
bugs as e.g. RelationshipSelectionCursor closed it's internal cursors
which handed them back to the transaction pool, meaning that they could
be handed back out on the next allocate even though we were not done
with them. Changing to the global non-pooling CursorFactory solves this,
while we still get the benefits of pooling from the cypher CursorPool.