Dynamic database addition/removal in discovery service
Every cluster member is responsible for updating the discovery
service when it starts/stops a database. Discovery service
keeps a set of started database IDs for every cluster member.
Before this change, the set of database IDs was only published
once for cores and periodically for read replicas. This means
cores published only default and system database IDs and read
replicas needed to wait for a refresh timeout to update the
discovery service.

This commit makes updating the set of database IDs in the discovery
service a pert of the database lifecycle. Core and read replica
databases will add themselves to the discovery service during start
and remove themselves during stop. Such functionality is necessary
when databases are created/started/stopped dynamically
in a neo4j cluster.