Do not access database manager from actors
MetadataActor and ClientTopologyActor publish information about
available databases for cores and read replicas respectively. Both
need to publish a set of started databases when an actor system
starts up. This set can be empty when a cluster is starting up
or non-empty when the actor system is restarted at runtime.

Before this change, both actors took the set of started databases
at runtime from DiscoveryMember, which accessed the database
manager directly. Access of such global mutable state is considered
an anti-pattern for actors.

This commit changes the implementation of DiscoveryMember to be
immutable. Discovery members are now created by a factory when
an actor system starts. This makes MetadataActor and
ClientTopologyActor access an immutable snapshot of started databases.