Fix flaky ClusterStateMigrationIT
Test shouldFailWhenClusterStateVersionIsWrong was flaky because
cluster startup is executed in multiple threads but the resulting
future fails when any of the combined futures fail. This could make
test directory extension try to clear directory of a cluster member
that hasn't been properly shutdown.

The problem is fixed by restarting only a single cluster member
with an incorrect cluster state version.