Handle cached properties after an aggregation
We where trying to use the original name that may not longer be in scope
for looking up slot configurations for the cached property. This broke
if you combined renamings and aggregations like
```
x.prop...WITH x AS y ... WITH collect(something) AS foo RETURN y.prop
```

We are then caching all properties under the name `x` but we cannot use that
to look up slot configurations for `y` since it is not in scope.