Fix reduce-apply-reduce in MorselArgumentStateBuffer
When initiating, we compute an Array of argument row ids that all the
reducers of an MorselArgumentStateBuffer care about for a particular
ArgumentState. This Array is put into the argument state, so that we
can decrement the rights argument row ids again when we close
the ArgumentState.

This same pattern needs to be replicated for other buffers.