Refactored ResourcePrivilege and others to use enum instead of boolean
The GRANT/DENY split was being passed around as a boolean, which was
less clear in the code what it meant, especially since the variable
was often the very generic `allowed`. The new approach uses specific
enum types GRANT and DENY. This also allows for simplification in many
places and the removal of some hard-coded strings for
GRANT, DEBY, GRANTED, DENIED, etc.