Invoke system DB transaction listeners for all transactions
Before this change, transaction event listeners were only invoked for
transactions that contained data changes. They were not invoked for
internal transactions that introduced tokens or schema changes. Such
behavior is problematic for system database listeners used to trigger
the reconciler.

Public API disallows listener registration for the system database.
Thus it is only possible to have DBMS-internal listeners for system
database transactions.

This commit makes the system database invoke its listeners on every
transaction. Such listeners are internal and trusted to handle a
possibly empty TransactionData. Public API behavior is unaffected
by this change.

It is possible for the listener to be invoked when a nested internal
transaction has committed. For example, token-introducing
transactions are nested and implicit. They can be committed but the
owning user transaction would remain open. There is no need to reconcile
such transactions. Thus SystemGraphDbmsOperator now checks if a
transaction in the caller thread is still open.

Do not reconcile when in transaction