Fixed so indexes don't give a restricted user values that is denied
When the user has GRANT and DENY on the same property for different
labels and there is an index on the allowed label, the index was used.
However, it returned the property value also on nodes with both
the allowed and the denied label where it should have been denied.