Wait for system database bookmarks based on reconciled IDs
Adapt TransactionIdTracker to wait for reconciled transactions on the
system database. It used to wait for committed transaction ID, which
is not correct for the system database. In clustered deployments
system operations are reconciled asynchronously and thus
TransactionIdTracker can't wait for committed transactions, as
it does for all other databases.

There are now two implementations of ReconciledTransactionTracker. One
for standalone community database that does not have a reconciler.
Another for multi-database aware standalone and clustered databases.
Tracker for standalone community databases only checks last committed
transaction IDs. Tracker for multi-database deployments has a separate
tracking of reconciled transaction IDs based on OutOfOrderSequence.