Avoid producing log messages about missing transaction state dependencies for the full text index provider when running in a TOOL context.
Tools, such as the consistency checker, do not create any transaction state infrastructure and that is perfectly fine, because they do not write or update any data.
They only read it, so there is no need for transaction state.
The logging is pretty harmless in the first place, but some customers still get nervous when they see messages like that, even when they are on debug log level.