Fix potential NPE in BetterRaftMessageLogger
This logger is started and stopped in a core database lifecycle.
However, it might sometimes be used by threads that do not
participate in this lifecycle and these threads can try to log
when the logger is stopped. Such threads are those used for election
and heartbeat timers. They are might not be fully stopped when a
RaftMachine is stopped because the timers are canceled in ASYNC mode
and these threads can remain running.

The problem is fixed by using a read-write lock in
BetterRaftMessageLogger to guard lifecycle methods and make sure
logging is a no-op when the logger is stopped.