Fix creation of AccessCapability in CC
Core databases need to create AccessCapability which knows about
the role of the cluster member. Only leader is allowed to write. This
commit changes Database to create AccessCapability using the provided
factory. In CC this factory comes from edition module and uses
RaftMachine to figure out the member role.

This fixes a few tests that expected WriteOperationsNotAllowedException
when writing on followers or read replicas.