Include error status in BarrierException
And propagate it through AcquireLockTimeoutException and
IdGenerationException that could be user-facing. It is important for
those exceptions to have a status because that's what is sent
back to the user by the Bolt server. Driver users receive error code
and error message. This commit also adjusts error messages in ID
generator and lock manager to be a bit friendlier.

Also, fixed TokenReplicationStressIT which now ignores all
"leader unavailable" types of errors.