Fuse Limit operator
Special case code generation template for limit when not
nested under an apply, and in single threaded execution
or in a serial pipeline in parallel execution, that
bypasses the argument state map filtering and uses
a simple counter and breaks out of the outer loop when
it reaches zero.

Adds instruction to break out of a labeled loop to the codegen library.