Removing Bolt V1 and Bolt V2
When tx failed with error, no more tx or query is allowed until the failure is cleared with a RESET message.
The client is not expected to continue a failed tx, as the tx can only be rolled back.
One of the main goal of this PR is to get rid of `ACK_FAILURE` message so that no more interaction is allowed with the tx after it is failed.