Mark panicked databases as failed in the reconciler
A panicked database should be marked as failed in the reconciler so
that it doesn't get transitioned to a different state. Transition of
a failed database is only possible when it is forced.

Also, fixed a problem in the reconciler where it did not keep the
database's failed state after no-op transitions.