Connect GBPTreeConsistencyChecker with GBPTreeConsistecyChecker
GBPTree use CleanTrackingConsistencyCheckVisitor internally to track
if any calls to the consistency check visitor was made instead of
every GBPTreeConsistencyCheckVisitor needing to implement it's own
tracking.

Introduce GBPTreeBootstrapper used to at least somewhat automatically
open a GBPTree on existing file. We still need to have some knowledge
about the expected layout of the index however... which is sad.
We use SchemaLayouts to do this.
The reason we cannot simple give a list of all known Layouts (like what
SchemaLayouts eventually will try to do) is because there is no way for
us currently to tell GenericLayout with 1 slot apart from a GenericLayout
with multiple slots. This need to be fixed.

Introduce ConsistencyCheckWithCorruptGBPTreeIT.