Makes use of highId from store when recreating id generator
So when the store exists, but the id generator doesn't it
will use the possibility to scan for high id in the store
when recreating the id generator at creation, even before
starting it (where rebuild happens). This because token and schema
stores are read after recovery, but before indexed id generators
have been started (in this case rebuilt).