Better error in backup when DB does not exist
Backup starts with a GetDatabaseIdRequest which resolves the specified
database name to a database ID. All subsequent requests use database ID.

This commit changes GetDatabaseIdRequestHandler to return an error
response instead of throwing when the specified database name is
unknown. Throwing an exception from the handler resulted in a channel
being closed and the client receiving a ClosedChannelException. Clients
receive proper error messages when an error response is returned.

Also, improved error messages in the backup command to include the
exception message. Before this commit, backup command printed only
"Execution of backup failed" when not in verbose mode.