Skip database_lock in restore command
Do not try to copy database_lock from the source to the target
directory. Do not try to remove database_lock from the target directory.

Restore command locks database_lock file in the target
directory and copying can result in AccessDeniedException on Windows.

This commit should fix RestoreDatabaseCommandIT and
ConvertNonCausalClusteringStoreIT on Windows.