Allow SET += / = Syntax on variables with no semantic information
SET x += {prop: 5} would previously fail if x had no semantic
information, even if it turned out to be a node/relationship at
runtime. This is now fixed.

This also fixes a tangential problem with invalidation of cached
properties in +=.

This also fixes a tangential problem with eagerization of SET for
different cases.