Reach for txState before allocating ID
This is because instantiation of txState is arguable the best
hook for knowing when a transaction becomes a write transaction.
This plays into the epoch tied to each transaction and which
causal clustering makes use of to ensure that transactions which
were started and had some write operation before a leader barrier
cannot commit after the barrier.