Changes EpochException to RuntimeException
This exception, which is sort of new, plays a role that
AcquireLockTimeoutException used to play. It's not a KernelException
because it needs to trickle out through the execution stack where
KernelException typically isn't present, but still needs to have
the properties of a KernelException, i.e. implementing HasStatus.
This is a sign of a problem where this sort of exception sits in
between two types of failure handling mechanisms to work around
some sort of design issue.

This change doesn't change any failure handling design,
neither does it make things worse. It mostly takes over some uses of
of AcquireLockTimeoutException to instead use EpochException, to signal
transient transaction failure issues, something that drivers
picks up on and retries the query. EpochException can additionally
be thrown from some tx lifecycle places.