Renames jar to neo4j-push-to-cloud-3.jar
This idea behind this jar naming is that the 3.x (Neo4j version)
jar(s) are all compatible and interchangeable with each other.
Also since the jar (some version of it) will be available as a
separately downloadable jar the user should not end up with
two of effectively the same jar file having different names
in the same installation.