Make it safe to stop a not started ClusterBindingHandler
CoreDatabaseLife could fail to start and then try to stop a not
started ClusterBindingHandler. This would cause an NPE because
ClusterBindingHandler does not have a RaftId. It would try to
de-register a handler chain from RaftMessageDispatcher using
null instead of a RaftId.

This commit fixes the problem by making ClusterBindingHandler
not de-register the handler chain if the handler has not been started.