Clear cluster properties from system graph when restoring cluster with seed
Previously these properties were added to the system graph to improve
the fault tolerance and availability for database creation. Specifically
the set of initial members were useful for that. However, when restoring
a new cluster based on a system database backup, this set of member ids
are no longer valid and prevent restoration of the previous set of
databases. The solution here is thus to remove those properties when
a new cluster is formed based on a system database seed.

All databases will thus be reformed in the new cluster using the
discovery method, instead of the initial members from system graph method.