Re-implement CartesianProduct to execute rhs once per lhs morsel...
This architecture transforms a cartesian product plan

         top
         CP
     lhs    rhs

into the execution graph

 lhs -> [ATTACH]-[APPLY]----------- ..... > [MrBuff LHS]
                        `[DELEGATE]--rhs- > [MrBuff RHS] -CP,top->

In this setup, the attach buffer fold each input morsel into an
attachment of a new 1-row morsel. The MrBuff LHS will be immediately
complete on being initiated with this new morsel, meaning that there
is no runtime pipeline between the apply and [MrBuff LHS]. The CP
operator take pairs of LHS and RHS morsels and streams the cartesian
product of these.