Same topology refresh interval on cores and RRs in tests
Before this commit, the interval was 1s on cores and 3s on RRs. This resulted
in a lot of unnecessary RR info pruning in core discovery services. Logs were
also quite confusing. Now cores and RRs use the same interval value, as it
should be in production. This makes logs more clear and removes the noise.