Fix ConsistencyCheckWithCorruptGBPTreeIT leaving files behind
Because NATIVE30 provider use Lucene internally we can not use the
snapshot from ephemeral file system because lucene will not use it to
store the files. The only options are to use a real persisted file
system or to use in memory variant in Lucene. When we do consistency
check we need to have the Lucene files around and we can only have that
if they are visible through our file system. Hence we use a the default
file system.