Fix issue with uninitialized locals
The way we track if the locals for slots has been seen already, and
thus does not need to be initialized from the execution context, is
not always sufficient if the code path of first access is in a
conditional branch.
This happened when we execute a continuation where innerLoop = true.

This is a sub-optimal fix that inserts a runtime check if a local
slot variable has not yet been initialized, on each use, until
we come up with a better solution.