Use pooled buffers in store copy
Previous solution did the following

FileSender
    - read part of file into fixed NIO direct byte buffer (max 8192)
    - allocate a new byte array on the heap
    - copy bytes from direct buffer into byte array
    - wrap byte array in a FileChunk object
    - send FileChunk down the netty pipe

FileChunkEncoder
    - allocate new direct netty buffer with size 256
    - copy from byte array in the FileChunk to netty buffer, growing as necessary
    - send netty buffer down the netty pipe

FileChunkDecoder
    - allocate new byte array on the heap
    - copy from netty buffer into byte array
    - wrap byte array in a file chunk object
    - send FileChunk up the netty pipe

FileChunkHandler
    - write payload from byte array on heap down into file

New solution does the following

FileSender
    - allocate a netty direct byte buffer (size 8192)
    - read part of file into netty buffer (max 8192)
    - wrap netty buffer in a FileChunk object
    - send FileChunk down the netty pipe

FileChunkEncoder
    - allocates a new netty direct byte buffer for the header (4)
    - combines the header and the payload in a composite buffer
    - sends the composite buffer down the netty pipe

FileChunkDecoder
    - retains the payload as a netty buffer slice
    - wrap payload in a FileChunk object
    - send FileChunk up the netty pipe

FileChunkHandler
    - write payload from netty direct buffer down into file

With the new solution
    - one allocation and two memory copies are removed on the outbound
    - one allocation and one memory copy is removed on the inbound
    - all buffers are generally pooled direct buffers (netty preference, but configurable)
    - garbage on the Java heap should be significantly reduced

The composite buffer is necessary because the protocol stack relies on framing
and a FileChunk must be encoded as a complete frame (one logical ByteBuf).

It would help performance to increase the payload size, but it is unfortunately
hardcoded into the protocol and would break backwards compatibility. This can be
fixed in the future with a redesign of the protocol and a new version. Another
future small performance improvement could be to remove the wrapping as FileChunk
objects.

Another area to take a look at in the future is transaction streaming, since this
change just deals with files.