Unflake QueryRestartIT
It has an intricate relationship with the page cache
version context feature. The test was over-sensitive to
anything marking any page as dirty from f.ex. index sampling
and id generator maintenance background jobs.

Now the test only cares about such calls if they come from
the thread executing the query. This could be troublesome to
keep if/when query execution goes parallel or when
query execution is handled differently, by means of having
a pool and just queue queries onto it, but that's a later problem.