Add folding analyzer
In accented languages it is common that users expect keywords
without diacritics to match text with diacritics and vice versa.

Lucene implements ASCIIFoldingFilter and ICUFoldingFilter for this
task. The latter works better on various unicode characters.
The ASCIIFoldingFilter is already on Neo4j classpath, while
ICUFoldingFilter is in another package (analyzers-icu), hence the
choice.
Closes $12305

Signed-off-by: Anton Klaren <anton.klaren@neotechnology.com>