Unflake RotatableCsvOutputIT
By making the MetricsTestHelper reading utility safer.
It used to wait for there being a file with > 1 number of rows in it.
After that wait it would open the file and read that (second) line.
In between may have been a rotation and so the reading would fail.
Now instead it will try to open the file, read header and then
one data line and if it succeeds it returns that result,
otherwise retries.