Make chunk size configurable in store copy and increase default from 8K to 32K
All protocols already use framing of message buffers so it is unnecesary to have
additional framing for file chunks. Removing the length field from the file chunk
allows it to be inferred from the outside framing instead and be flexible. The
chunk size can thus now be a configurable value.

The increase of the default is based on some very basic manual benchmarking,
since we don't have proper benchmarks. This causes a network protocol change
but 4.0 is already breaking in that regard, so it is fine.

The FileChunk wrapper can almost be entirely removed now, but I'll leave that
as follow up work for the future when time permits.