Fixes txId snapshot issue in SequenceArray
SequenceArray could sometimes miss some of the out-of-order
ids it was offered when doing a snapshot, which would have the
counts store re-apply the same transaction during recovery and
so produce the wrong count in the end.
Also unflakes GBPTreeCountsStoreTest in the process.
Made the test way more likely to trigger the snapshot issue,
basically 100% w/o the fix in SequenceArray. This also enabled
a change to make the test time shorter.