Add support for whitelisting dynamic configuration values
Introduce whitelisting of properties that can be configured dynamically.
By default, all availability properties will be whitelisted (*).
Server administrators can, however, override this setting to whitelist
only specific properties.

If this configuration parameter is set to the empty string, all dynamic
configuration is turned off.

If a user with the admin role calls `dbms.setConfigValue` on property
that has not been whitelisted, error will be thrown.