Smaller ID caches are used more often than large ones
Now the IndexedIdGenerator constructor accepts one more argument
which must be set to true to allow large ID caches, otherwise
a smaller one is used. This to generally lower the memory usage
of the ID caching. The ability to inject this from the outside
allows use to use small caches for the system database.

Also an internal setting `unsupported.dbms.force_small_id_cache`
was introduced to let specific use cases (I'd guess mostly tests)
to force small caches even for non-system databases. This allowed
ReuseStorageSpaceIT to set this setting and to therefore be able
to work with much smaller data sizes and still exercise the ID reuse
and caching equally much as before. The test time should go from
~12 min down to ~1 min.