Enhance transaction logs infrastructure to support files in different formats
Recent updates to transaction logs format bump size of the file header.
This means that now we can have log files in different formats (3.5 and 4.0)
that we should be able to read and extract info from.
This PR removes assumptions that initial data-position are the same in all log files
and add the possibility to extract any format header information from any log file.