Remove --database from unbind command
Unbind command is a DBMS-level one and does not work on a specific database.
It used the provided database name to verify that this database exists but
then deleted all of the cluster state for all Raft groups.

This commit removes `--database` command line parameter from `neo4j-admin unbind`.
It also makes the command check DBMS lock file (`store_lock`) instead of a
database-specific lock file (`database_lock`) of the specified database.