Make sure expand cursor can be reused
In a previous optimization we made sure we didn't need to allocate
group cursor for non-dense nodes. In doing so we weren't reusing the
traversal and group cursors any more since they are hidden in
the SelectionCursor and nulled on close. This PR make sure we don't do
the extra allocation of group cursors while still making sure we can
reuse the curors.