Coordinating leases for transactions with unknown fates
Short description

When the replicator gives up trying to replicate a transaction
before it has been applied but after a replication attempt has
been made, then the locks held will be released, but since
they are locking entities that might still get updated, every
other transaction working under the same locking lease must
be aborted or they would risk operating on stale data.

Longer descriptions

- Change terminology used from barrier token / epoch to lease

Leases are a more well-known concept in distributed systems and it fits
neatly with what we have done here as well. Eliminating the use of the
token term also removes the confusion potential it has with kernel tokens.

- Invalidate lease locally when transaction aborts early

Leases are used for coordinating the validity of locks which operate
against local lock managers. Transactions that are aborted locally
release their local locks, but if they've already entered the replication
pipeline their effects might still eventually take effect, meaning
they may have not been fully aborted. However, this breaks the concept of
locking since there might be other transactions blocked by a lock under
the same lease which get released before these effects are completed.

The solution is to invalidate the lease locally when such a transaction
with an unclear fate aborts. This is implemented as an interplay between
the ClusterLeaseCoordinator and the ReplicatedTransactionCommitProcess.
Transactions that are blocked behind a transaction with an unclear fate
will be guaranteed to also abort since the lease will have been invalidated
before they reach the ReplicatedTransactionCommitProcess.

Note that there still is only a single valid lease at any point in
time as ordered by the Raft for the particular database.

- Introduce a ReplicationResult to capture the possibilities better

The result of a replication attempt can either be that it was:
  * known to not have been replicated at all (e.g. there was no leader initially)
  * unknown whether or not it was replicated (e.g. leader lost its role,
    and replicator stopped retrying, but an attempt was made and might
    still take effect)
  * replication completed fully and its effects are locally visible
    (e.g. replicated transaction was locally applied to the store)

The novelty in this scheme is really that of exposing to users of the
replicator the fact that a replication attempt was aborted but still
might take effect.

The throwing of exceptions for reporting results from the replicator
has been removed, but any underlying failures are still captured as
exceptions in the ReplicationResult.

- Renamed Result to StateMachineResult

This makes it clearer in code that this is a result of a replicated
command having been applied to a state machine.

- Remove NoLeaderAvailable status code

The driver does not care about that particular code and throwing
NotALeader is more appropriate since it will cause the driver
to remove us as a writer from the routing table.