Don't do `break` from inner loops and fix profiling
When fusing-over-pipelines inner-loops would do a `break` when there were
more stuff to do but no more data could be written. The reason for the
`break` was really just to avoid advancing the input-morsel and avancing
cursors. However this caused issues where operator state code wasn't
being run. Here we change so that we don't do any `break` and instead
conditionally advance input-morsel, cursors and such.

This commit also applies a separate fix for profiling to avoid
double-counting rows after a continuation.