Fix flaky ConnectionTrackingIT
A static instance of a Java HttpClient used in HTTP class was most likely
sending multiple HTTP requests through the same connection. This made the
test flaky because it expected a single request to result in a single connection.

This commit fixes the problem by making the test use a new HTTP client for every
request. HTTP clients are also kept in a set until the test completes. This is
because HTTP clients are reference-counted and can be terminated while a test
is running by the GC. Termination of a HTTP client closes its connection and
stops the selector thread.