Rename tx-logs to transactions to avoid confusion with log files.
The fact that it's called write ahead LOG is an inconvenient coincident, that have ended with users deleting all the log files when they run out of disk. This rename will remove the word LOG from all the files that are crucial to the database consistency to avoid this scenario.