Fix a lifecycle race in RotatableCsvReporter
This reporter holds a map of all open file writers. It also has a separate
thread that periodically reports metrics and writes to CSV files. File
writers are lazily added to the map by the reporter thread. All writers
are closed when the RotatableCsvReporter is stopped.

Stopping the RotatableCsvReporter could happen concurrently with reporting
new metrics. This could result in a writer added to the ConcurrentHashMap
of writers concurrently with #stop() iterating over this map. Such a writer
might not get seen by #stop() and might not get closed. This resulted in ITs
failing sometimes because a test directory can't be cleaned up:

```
java.nio.file.FileSystemException: C:\...\neo4j.bolt.accumulated_queue_time.csv:
The process cannot access the file because it is being used by another process.
```

The problem is fixed by synchronizing stop and report operations. This is consistent
with ScheduledReporter which RotatableCsvReporter extends. ScheduledReporter
synchronizes its #stop() and #report() but we override #stop().