Add checksum to transaction logs
Checksums are calculated over each transaction so we can validate the integrity of the stored files when doing recovery. CRC32C was chosen because it should theoretically give better performance when working in buffered reader/writer as we do.

Also, remove masterId and authorId from transactions since that is no longer the source for checksums.