Handle index seek of multiple boolean properties in compiled runtime
In the compiled runtime we are internally using an `IntStream` to store
lists of booleans. This caused an issue when doing an index seek for
multiple index properties since values weren't being properly converted
back to booleans before seeking. So instead of seeking for `true` we
looked for `1` and instead of `false` we seeked `0` leading to the wrong
results.

Fixes $12328

changelog: Fix bug for index queries with multiple boolean properties,
like `MATCH (n:L) WHERE n.p = true OR n.p = false RETURN n.p`