listTransaction now lists committing transactions
* listTransaction now lists committing transactions

The transactions go through the following states, Open -> Closing -> Closed. The closing state is where commit or rollback happens, and the transaction can be stuck there. Before, `listTransactions` did not show transactions in the closing state, hiding the fact that a transaction was still ongoing and led to some confusion.

* Added error message when trying to kill closing transactions.
* Remove setting `dbms.procedures.kill_query_verbose` since it was useless.
* Fewer streams, it's good for you...