Unflake CoreFallBehindIT
Formation reduction was just temporarily needed until we supported
forming databases with less than full amount of members. Now it
can be a source of cluster formation split, so removing it.

The runtime setting still needs to be lower to allow for the removal
and subsequent detection of a member being removed from the system
raft group, upon which the test relies.

An extra assertion was added to await that the system database is
indeed fully formed before disconnecting the chosen member.