Add utility for accessing cursors directly from expressions
For queries such as `MATCH ()-[r]->() RETURN type(r)` we used to use a
separate cursor for fining out the type of `r`. This means we will need
to do

```
read.singleRelationship(id, cursor)
```

to position the cursor correctly. However there is - for this query - a
cursor in use that points to the correct location. This commits introduce
a mechanism to leverage this when evaluating expression.