Add test helper Assert#awaitUntilAsserted()
This new method will execute the given block of code until it doesn't
throw an AssertionError. It makes it easy to convert blocks of code
that use JUnit asserts into eventual assertions. The functionality is
very similar to Assert#assertEventually() but does not require
rewriting all assertions into a Hamcrest matcher.