Let GBPTreeConsistencyCheckVisitor take care of 'isDirty' check
Instead of having dedicated consistency check tasks and isDirty() method
through whole stack.

The isDirty inconsistencies will be logged as a warning, just like it
was before.

If a btree index isDirty it will be excluded from further consistency
checks just like with any other structural inconsistency, even if the
inconsistency itself will be logged as a warning.