Fixes an initialization issue with restartable importer
Where, if am import would crash in the middle of creating the NeoStores
The empty values would then later on be mis-interpreted as real values
and only when later creating the transaction log file
(which would have version number -1) and starting the database
the metadata store fields would be seen as uninitialized and fail.
Now the test detects this itself such that an import that crashes
while creating the store files will have those store files deleted
and recreated again on the next attempt.